StoreSubscriptionCheckService had been refactored
